Add files via upload
Currently, the program could add vertices and edges into a graph. I tested some specific cases. All cases succeeded and correctly added edges into the graph.

In the main.cpp, there is a commented function which is named as "addEdges". This function is able to add edges. We could notice that there are a lot of duplicate codes in the main function. These codes can be replaced by "addEdges" function making the program clearer and shorter. However, passing pointers and passing references will take up some memory when the program is calling it. I have not decided whether to replace these duplicate codes by the "addEdges" function.